[#32] Add/remove object to queues when flagging or retracting.

+++ routes/flag/delete.php
@@ -7,7 +7,10 @@ $userId = User::getActiveId();
 
 // Skip error checking. If the user somehow flagged the object, the user
 // should be able to unflag.
-Flag::delete_all_by_userId_objectType_objectId($userId, $objectType, $objectId);
+if ($userId) {
+  Flag::delete_all_by_userId_objectType_objectId($userId, $objectType, $objectId);
+  Queue::check($objectType, $objectId);
+}
 
 header('Content-Type: application/json');
 print json_encode(_('Your flag was deleted.'));
